refactor: getFeedInfo 메서드 FeedService 로 분리
