+++ src/main/java/com/woowacourse/dsgram/service/ArticleService.java
@@ -9,11 +9,9 @@ import com.woowacourse.dsgram.domain.repository.CommentRepository;
 import com.woowacourse.dsgram.domain.repository.LikeRelationRepository;
 import com.woowacourse.dsgram.service.assembler.ArticleAssembler;
 import com.woowacourse.dsgram.service.assembler.UserAssembler;
-import com.woowacourse.dsgram.service.dto.FeedInfo;
 import com.woowacourse.dsgram.service.dto.article.ArticleEditRequest;
 import com.woowacourse.dsgram.service.dto.article.ArticleInfo;
 import com.woowacourse.dsgram.service.dto.article.ArticleRequest;
-import com.woowacourse.dsgram.service.dto.follow.FollowRelation;
 import com.woowacourse.dsgram.service.dto.user.LoggedInUser;
 import com.woowacourse.dsgram.service.dto.user.UserInfo;
 import com.woowacourse.dsgram.service.strategy.ArticleFileNamingStrategy;
@@ -141,23 +139,6 @@ public class ArticleService {
         return articles.map(
                 article -> findArticleInfo(article.getId(), viewerId)
         );
-    }
-
-    public FeedInfo getFeedInfo(String fromNickName, String toNickName) {
-        User guest = userService.findByNickName(fromNickName);
-        User feedOwner = userService.findByNickName(toNickName);
-        long followers = followService.getCountOfFollowers(feedOwner);
-        long followings = followService.getCountOfFollowings(feedOwner);
-        List<Article> articles = findArticlesByAuthorNickName(toNickName);
-        FollowRelation followRelation = followService.isFollowed(guest, feedOwner);
-
-        return FeedInfo.builder()
-                .user(feedOwner)
-                .followers(followers)
-                .followings(followings)
-                .articles(articles)
-                .followRelation(followRelation)
-                .build();
     }
 
     @Transactional
