+++ src/main/java/com/woowacourse/dsgram/service/dto/FeedInfo.java
@@ -1,11 +1,8 @@
 package com.woowacourse.dsgram.service.dto;
 
-import com.woowacourse.dsgram.domain.Article;
 import com.woowacourse.dsgram.domain.User;
 import com.woowacourse.dsgram.service.dto.follow.FollowRelation;
 import lombok.*;
-
-import java.util.List;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
@@ -15,15 +12,13 @@ public class FeedInfo {
     private User user;
     private long followers;
     private long followings;
-    private List<Article> articles;
     private FollowRelation followRelation;
 
     @Builder
-    public FeedInfo(User user, long followers, long followings, List<Article> articles, FollowRelation followRelation) {
+    public FeedInfo(User user, long followers, long followings, FollowRelation followRelation) {
         this.user = user;
         this.followers = followers;
         this.followings = followings;
-        this.articles = articles;
         this.followRelation = followRelation;
     }
 }
