+++ src/main/resources/static/js/template.js
@@ -103,7 +103,7 @@ const TEMPLATE_APP = (() => {
                                             좋아합니다.
                                         </p>
                                     </div>
-                                    <div id="'article-contents-${articleInfo.articleId}" class="feed-contents pdd-left-15">
+                                    <div id="article-contents-${articleInfo.articleId}" class="feed-contents pdd-left-15">
                                         <p class="contents">${articleInfo.contents}</p>
                                         <form style="display: none;">
                                             <input type="text" value="${articleInfo.contents}">
